Version extension moved from 1.0 to 1.1.8
We start giving the extension (i.e. what created with create extension)
the same version number of pg_reorg. This doesn't mean they will
always stay the same: if there is a fix to the C library but not
to the sql the extension version should not be increased (otherwise
it would take a string of empty files like 1.1.8--1.1.9.sql to create
an upgrade path).

The current version number is not hardcoded anymore in any file or file
name, but is read by the makefile from the .control file. The variable is
not used by the unpackaged upgrade file, as there should be no need to
release more of these.