Allow RDS customers to use the pg_repack client.
The current client checks for superuser before attempting to
execute pg_repack commands.  In Amazon RDS, customers are given
access to a psuedo-superuser role called rds_superuser, but they
are not given access to superuser.  However, rds_superusers will
otherwise have the ability to execute pg_repack commands in RDS.

This change alters the client code to ensure that roles with
membership to the rds_superuser role pass the superuser checks.