Use plain "create table as" if possible
We need to split the copying of data into the new table into separate commands if we need any per-column storage settings.
In this case we must use this order:
- create table as .. with no data;
- alter table .. alter column ..
- insert into .. select

But this is pointless if we don't need to change the column.
At the same time, a single "create table as" command may be very useful for performing --order-by because it may use parallel query execution feature in PostgreSQL. INSERT .. SELECT currently cannot use parallel plans.