Use savepoints when retrying to take AccessShareLock.
In lock_access_share() function, before acquiring an AccessShareLock
on the target table, concurrent DDL commands are terminated
unconditionally. However, there is a window between this termination
and the actual "LOCK TABLE ... IN ACCESS SHARE MODE" operation,
allowing other DDL commands to run on the target table. If the LOCK
TABLE failed due to a lock conflict, the transaction was rolled back
and retried to execute LOCK TABLE, but the LOCK TABLE command could
not be executed again as the transaction had already closed.

This commit fixes the issue by using SAVEPOINT and ROLLBACK TO
SAVEPOINT so that we can retry to take a lock in the same top
transaction.

Issue #383