Grant required perms except CREATE to PUBLIC for non-superuser usage
This covers all grants except for CREATE on the repack schema to allow
table owners to use pg_repack with --no-superuser-check. Importantly,
table ownership permissions still prevent users from accessing
intermediate repack data and log tables of other users, or attempting to
repack tables they do not own - similar to VACUUM FULL.

CREATE on the repack schema is not included to match the default
behavior of postgres v15+ of not allowing new roles to create tables
without an explicit grant.

We also add a successful non-superuser test case by replacing the repack
schema permission grants with a test that creates a user-owned table and
verifies pg_repack works for non-superuser with --no-superuser-check
flag.